refactor: reorganize DuckDB extension baking process and update scripts
- Removed the inline baking command from the Dockerfile and replaced it with a copy of the baked DuckDB extensions from the builder stage, ensuring a single baking mechanism across local and CI builds.
- Updated the `package.json` to remove the `bake-duckdb-extensions` command from the main build process, streamlining the build steps.
- Introduced a new `bake-duckdb-extensions.js` script in the server package to handle the baking of DuckDB extensions, improving reliability and consistency in environments with restricted internet access.
- Adjusted the `.dockerignore` to exclude the scripts directory while ensuring necessary scripts are included in the Docker build.

This update enhances the management of DuckDB extensions, ensuring they are available at runtime without relying on network access during the build process.
