+++ packages/server/package.json
@@ -17,8 +17,9 @@
       "test": "bun run test:unit && bun run test:integration",
       "test:unit": "bun test --timeout 100000 src",
       "test:integration": "bun test --timeout 200000 tests --max-workers=1",
-      "build": "bun generate-api-types && bun build:app && NODE_ENV=production bun run build.ts",
-      "build:server-only": "bun generate-api-types && NODE_ENV=production bun run build.ts",
+      "build": "bun generate-api-types && bun build:app && NODE_ENV=production bun run build.ts && bun run bake-duckdb-extensions",
+      "build:server-only": "bun generate-api-types && NODE_ENV=production bun run build.ts && bun run bake-duckdb-extensions",
+      "bake-duckdb-extensions": "bun scripts/bake-duckdb-extensions.js",
       "start": "NODE_ENV=production bun run ./dist/server.mjs",
       "start:init": "NODE_ENV=production bun run ./dist/server.mjs --init",
       "start:dev": "NODE_ENV=development bun --watch src/server.ts",
